+++ src/Alpakka.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Amqp.V1.Tests"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Amqp", "Amqp", "{9A8B3B60-F75D-4906-B368-739CA08C25D4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Azure.EventHub", "Azure\Akka.Streams.Azure.EventHub\Akka.Streams.Azure.EventHub.csproj", "{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Azure", "Azure", "{76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Azure.ServiceBus", "Azure\Akka.Streams.Azure.ServiceBus\Akka.Streams.Azure.ServiceBus.csproj", "{F0251C77-28A3-4815-919C-65E4823DC0D0}"
@@ -43,8 +41,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Xml.Tests", "Xml\Akka.Streams.Xml.Tests\Akka.Streams.Xml.Tests.csproj", "{BCE118B4-AAB2-4441-9EA0-45008FA14B91}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{00D302CC-A29C-47D2-B2EF-F236656A2608}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Azure.EventHub.Examples", "Azure\Examples\Akka.Streams.Azure.EventHub.Examples\Akka.Streams.Azure.EventHub.Examples.csproj", "{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.Azure.ServiceBus.Examples", "Azure\Examples\Akka.Streams.Azure.ServiceBus.Examples\Akka.Streams.Azure.ServiceBus.Examples.csproj", "{8C13FB84-16A1-40F5-9ECD-90D7D99D9901}"
 EndProject
@@ -100,9 +96,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.File", "File\A
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akka.Streams.File.Tests", "File\Akka.Streams.File.Tests\Akka.Streams.File.Tests.csproj", "{D3217296-C1D3-4156-9D88-2752FFCB582C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Streams.Azure.EventHub.V5", "Azure\Akka.Streams.Azure.EventHub.V5\Akka.Streams.Azure.EventHub.V5.csproj", "{FB973142-2F74-415E-B588-ADDCBA5F9F99}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Streams.Azure.EventHub", "Azure\Akka.Streams.Azure.EventHub\Akka.Streams.Azure.EventHub.csproj", "{FB973142-2F74-415E-B588-ADDCBA5F9F99}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Streams.Azure.EventHub.V5.Examples", "Azure\Examples\Akka.Streams.Azure.EventHub.V5.Examples\Akka.Streams.Azure.EventHub.V5.Examples.csproj", "{FA1BD6C6-8EDE-4671-A11E-0A7A28B49F71}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Streams.Azure.EventHub.Examples", "Azure\Examples\Akka.Streams.Azure.EventHub.Examples\Akka.Streams.Azure.EventHub.Examples.csproj", "{FA1BD6C6-8EDE-4671-A11E-0A7A28B49F71}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SignalR", "SignalR", "{EF448678-4308-4DA8-9EF6-3FC16E9571E1}"
 EndProject
@@ -126,10 +122,6 @@ Global
 		{8D0DED3F-28E9-4140-9765-FB3B9C8B3AAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D0DED3F-28E9-4140-9765-FB3B9C8B3AAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D0DED3F-28E9-4140-9765-FB3B9C8B3AAE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F0251C77-28A3-4815-919C-65E4823DC0D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F0251C77-28A3-4815-919C-65E4823DC0D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F0251C77-28A3-4815-919C-65E4823DC0D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -158,10 +150,6 @@ Global
 		{BCE118B4-AAB2-4441-9EA0-45008FA14B91}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BCE118B4-AAB2-4441-9EA0-45008FA14B91}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BCE118B4-AAB2-4441-9EA0-45008FA14B91}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8C13FB84-16A1-40F5-9ECD-90D7D99D9901}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8C13FB84-16A1-40F5-9ECD-90D7D99D9901}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8C13FB84-16A1-40F5-9ECD-90D7D99D9901}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -241,7 +229,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{209333C3-0916-47AE-B40C-70548E46A3BC} = {9A8B3B60-F75D-4906-B368-739CA08C25D4}
 		{8D0DED3F-28E9-4140-9765-FB3B9C8B3AAE} = {9A8B3B60-F75D-4906-B368-739CA08C25D4}
-		{5AEAB925-5BAB-4275-BAA0-2C2BFDC65212} = {76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}
 		{F0251C77-28A3-4815-919C-65E4823DC0D0} = {76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}
 		{C088D6C4-C4E1-4A07-888B-D449295F470B} = {76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}
 		{2BD96ED1-2DD5-4DA8-B5DF-569C30828E51} = {76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}
@@ -251,7 +238,6 @@ Global
 		{E7028A0A-5F18-4D38-B4BE-D86760ABD028} = {9D36CAB6-44DB-4390-9455-AFCDFCB8476E}
 		{BCE118B4-AAB2-4441-9EA0-45008FA14B91} = {9D36CAB6-44DB-4390-9455-AFCDFCB8476E}
 		{00D302CC-A29C-47D2-B2EF-F236656A2608} = {76BFA9AB-B23C-4B42-8FCC-0CEBED2CBFA9}
-		{8CC9CADD-1DAA-49D8-8BF9-1E3CC93FBC0C} = {00D302CC-A29C-47D2-B2EF-F236656A2608}
 		{8C13FB84-16A1-40F5-9ECD-90D7D99D9901} = {00D302CC-A29C-47D2-B2EF-F236656A2608}
 		{226204F3-CE6B-4912-943C-28DED48DD3C5} = {D16AA3B8-C8FD-4954-9123-039A2A0BEA19}
 		{6FAA514D-8B69-4EA7-9D17-F542C8B66ECF} = {D16AA3B8-C8FD-4954-9123-039A2A0BEA19}
