--- src/Azure/Akka.Streams.Azure.EventHub.V5/Akka.Streams.Azure.EventHub.V5.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <AssemblyTitle>Akka.Streams.Azure.EventHubs.V5</AssemblyTitle>
-        <Description>Azure EventHubs adapter for Akka.NET Streams</Description>
-        <TargetFrameworks>$(NetStandardLibVersion);$(NetLibVersion)</TargetFrameworks>
-        <PackageTags>$(AkkaPackageTags);azure;EventHubs</PackageTags>
-        <RootNamespace>Akka.Streams.Azure.EventHub</RootNamespace>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Akka.Streams" Version="$(AkkaVersion)" />
-        <PackageReference Include="Azure.Messaging.EventHubs.Processor" Version="5.11.3" />
-    </ItemGroup>
-
-</Project>
