+++ src/Azure/Akka.Streams.Azure.EventHub/Argument.cs
