+++ src/Azure/Akka.Streams.Azure.EventHub/BatchedEventProcessorClient.cs
