+++ src/Azure/Akka.Streams.Azure.EventHub/EventHubSink.cs
@@ -1,12 +1,21 @@
-﻿using System;
+﻿//-----------------------------------------------------------------------
+// <copyright file="EventHubSink.cs" company="Akka.NET Project">
+//     Copyright (C) 2013-2022 .NET Foundation <https://github.com/akkadotnet/akka.net>
+// </copyright>
+//-----------------------------------------------------------------------
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Akka.Streams.Azure.Utils;
+using Akka.Actor;
+using Akka.Event;
 using Akka.Streams.Dsl;
 using Akka.Streams.Stage;
-using Akka.Streams.Supervision;
-using Microsoft.Azure.EventHubs;
+using Azure.Messaging.EventHubs;
+using Azure.Messaging.EventHubs.Producer;
+using Decider = Akka.Streams.Supervision.Decider;
+using Directive = Akka.Streams.Supervision.Directive;
 
 namespace Akka.Streams.Azure.EventHub
 {
@@ -22,17 +31,18 @@ namespace Akka.Streams.Azure.EventHub
             private readonly EventHubSink _sink;
             private readonly TaskCompletionSource<NotUsed> _completion;
             private readonly Decider _decider;
-            private Action<Tuple<Task, List<EventData>>> _eventsSend;
             private bool _isSendInProgress;
+            private IActorRef? _self;
+            private readonly List<EventData> _pendingSend = new List<EventData>();
 
             public Logic(EventHubSink sink, Attributes inheritedAttributes, TaskCompletionSource<NotUsed> completion) : base(sink.Shape)
             {
                 _sink = sink;
                 _completion = completion;
                 _decider = inheritedAttributes.GetDeciderOrDefault();
-
+                
                 SetHandler(sink.In,
-                    onPush: () => TrySend(Grab(sink.In).ToList()),
+                    onPush: () => TrySend(Grab(sink.In).ToArray()),
                     onUpstreamFinish: () =>
                     {
                         // It is most likely that we receive the finish event before the task from the last element has finished
@@ -50,49 +60,85 @@ namespace Akka.Streams.Azure.EventHub
 
             public override void PreStart()
             {
+                _self = GetStageActor(args =>
+                {
+                    var (_, msg) = args;
+                    switch (msg)
+                    {
+                        case Status.Success _:
+                            if (_pendingSend.Count > 0)
+                            {
+                                var events = _pendingSend.ToArray();
+                                _pendingSend.Clear();
+                                TrySend(events);
+                            }
+                            else
+                            {
+                                _isSendInProgress = false;
+                                PullOrComplete();
+                            }
+                            break;
+                        case Status.Failure fail:
+                            _isSendInProgress = false;
+                            var cause = fail.Cause;
+                            switch (_decider(cause))
+                            {
+                                case Directive.Stop:
+                                    // Throw
+                                    _completion.TrySetException(cause);
+                                    FailStage(cause);
+                                    break;
+                                case Directive.Resume:
+                                    // Take the next element or complete
+                                    PullOrComplete();
+                                    break;
+                                case Directive.Restart:
+                                    // Try again
+                                    TrySend((EventData[])fail.State);
+                                    break;
+                                case var unknown:
+                                    throw new ArgumentException(
+                                        $"Invalid failure directive {unknown}. Valid values are " +
+                                        $"{nameof(Directive.Stop)}, {nameof(Directive.Restart)}, " +
+                                        $"and {nameof(Directive.Resume)}");
+                            }
+                            break;
+                        default:
+                            Log.Warning($"Unhandled EventHubSink message: [{msg.GetType()}]");
+                            break;
+                    }
+                }).Ref;
+
                 // Keep going even if the upstream has finished so that we can process the task from the last element
                 SetKeepGoing(true);
-                _eventsSend = GetAsyncCallback<Tuple<Task, List<EventData>>>(OnEventsSend);
                 // Request the first element
                 Pull(_sink.In);
             }
 
-            private void TrySend(List<EventData> events)
+            private void TrySend(EventData[] events)
             {
                 _isSendInProgress = true;
-                _sink._client.SendAsync(events).ContinueWith(t => _eventsSend(Tuple.Create(t, events)));
+                var localEvents = events;
+                SendBatch(events: localEvents)
+                    .PipeTo(
+                        recipient: _self,
+                        sender: _self, 
+                        failure: ex => new Status.Failure(ex, localEvents));
             }
 
-            private void OnEventsSend(Tuple<Task, List<EventData>> t)
+            private async Task<Status> SendBatch(IEnumerable<EventData> events)
             {
-                _isSendInProgress = false;
-                var task = t.Item1;
-                var events = t.Item2;
-
-                if (task.IsFaulted || task.IsCanceled)
+                var batch = await _sink._client.CreateBatchAsync();
+                foreach (var eventData in events)
                 {
-                    switch (_decider(task.Exception))
+                    if(!batch.TryAdd(eventData))
                     {
-                        case Directive.Stop:
-                            // Throw
-                            _completion.TrySetException(task.Exception);
-                            FailStage(task.Exception);
-                            break;
-                        case Directive.Resume:
-                            // Try again
-                            TrySend(events);
-                            break;
-                        case Directive.Restart:
-                            // Take the next element or complete
-                            PullOrComplete();
-                            break;
-                        default:
-                            throw new ArgumentOutOfRangeException();
+                        _pendingSend.Add(eventData);
                     }
                 }
-                else
-                    PullOrComplete();
 
+                await _sink._client.SendAsync(batch);
+                return Status.Success.Instance;
             }
 
             private void PullOrComplete()
@@ -115,20 +161,20 @@ namespace Akka.Streams.Azure.EventHub
         /// <summary>
         /// Creates a <see cref="Sink{TIn,TMat}"/> for the Azure EventHub
         /// </summary>
-        /// <param name="client">The <see cref="EventHubClient"/> that sends the events to the EventHub</param>
+        /// <param name="producer">The <see cref="EventHubProducerClient"/> that sends the events to the EventHub</param>
         /// <returns>The <see cref="Sink{TIn,TMat}"/> for the Azure EventHub</returns>
-        public static Sink<IEnumerable<EventData>, Task> Create(EventHubClient client)
+        public static Sink<IEnumerable<EventData>, Task> Create(EventHubProducerClient producer)
         {
-            return Sink.FromGraph(new EventHubSink(client));
+            return Sink.FromGraph(new EventHubSink(producer));
         }
 
-        private readonly EventHubClient _client;
+        private readonly EventHubProducerClient _client;
 
         /// <summary>
         /// Create a new instance of the <see cref="EventHubSink"/>
         /// </summary>
-        /// <param name="client">The <see cref="EventHubClient"/> that sends the events to the EventHub</param>
-        public EventHubSink(EventHubClient client)
+        /// <param name="client">The <see cref="EventHubProducerClient"/> that sends the events to the EventHub</param>
+        public EventHubSink(EventHubProducerClient client)
         {
             _client = client;
             Shape = new SinkShape<IEnumerable<EventData>>(In);
