+++ src/Azure/Akka.Streams.Azure.EventHub/Extensions.cs
