+++ src/Azure/Akka.Streams.Azure.EventHub/ProcessEventBatchArgs.cs
