+++ src/Azure/Akka.Streams.Azure.EventHub/ProcessorFactory.cs
