+++ src/Azure/Examples/Akka.Streams.Azure.EventHub.Examples/Akka.Streams.Azure.EventHub.Examples.csproj
@@ -1,8 +1,10 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
         <TargetFramework>$(NetTestVersion)</TargetFramework>
+        <Nullable>enable</Nullable>
+        <RootNamespace>Akka.Streams.Azure.EventHub.Examples</RootNamespace>
         <IsPackable>false</IsPackable>
     </PropertyGroup>
