+++ src/Azure/Examples/Akka.Streams.Azure.EventHub.Examples/Multiple/StreamProcessorActor.cs
