+++ src/Azure/Examples/Akka.Streams.Azure.EventHub.Examples/ProcessorFactory.cs
