+++ src/Azure/Examples/Akka.Streams.Azure.EventHub.Examples/Single/SingleProcessorExample.cs
