--- src/Azure/Examples/Akka.Streams.Azure.EventHub.V5.Examples/Akka.Streams.Azure.EventHub.V5.Examples.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>$(NetTestVersion)</TargetFramework>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Akka.Streams.Azure.EventHub.Examples</RootNamespace>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\Akka.Streams.Azure.EventHub.V5\Akka.Streams.Azure.EventHub.V5.csproj" />
-    </ItemGroup>
-
-</Project>
